requirements.txt
bcrypt~=4.0.1
PyJWT~=2.8.0
Flask~=2.2.3
mongoengine~=0.27.0
pymongo~=4.4.1
boto3~=1.28.20
botocore~=1.31.20
flask-mongoengine
Flask-Cors